More simple optimizations, and a better UI design. More customization added to hacks.

+++ src/minecraft/net/minecraft/entity/EntityLivingBase.java
@@ -853,7 +853,7 @@ public abstract class EntityLivingBase extends Entity
 
     public void setHealth(float health)
     {
-        dataWatcher.updateObject(6, Float.valueOf(MathHelper.clamp_float(health, 0.0F, getMaxHealth())));
+        dataWatcher.updateObject(6, MathHelper.clamp_float(health, 0.0F, getMaxHealth()));
     }
 
     /**
